refactor: reorganize auth code and remove barrel exports
- Move all auth card components, shared components, and hooks to src/components/sign_up
- Rename page-level views in src/views/sign_up to .entrypoint.tsx for consistency
- Remove unnecessary src/views/sign_up/components/index.ts barrel export
- Update all imports to use direct file paths
