+++ src/components/sign_up/AuthPageLayout.tsx
