+++ src/components/sign_up/SignUpCard.tsx
@@ -4,16 +4,16 @@ import { useNavigate } from 'react-router-dom';
 
 import { BRAND_PRIMARY } from '../../design';
 import { supabase } from '../../utils/supabase';
-import { Paths } from '../paths';
+import { Paths } from '../../views/paths';
+import AuthCard from './AuthCard';
+import AuthCardHeader from './AuthCardHeader';
 import {
-  AuthCard,
-  AuthCardHeader,
-  AuthSubmitButton,
-  AuthSuccessMessage,
   ConfirmPasswordField,
   EmailField,
   PasswordField,
-} from './components';
+} from './AuthFormFields';
+import AuthSubmitButton from './AuthSubmitButton';
+import AuthSuccessMessage from './AuthSuccessMessage';
 import useAuthAction from './hooks/useAuthAction';
 
 type TSignUpForm = {
