+++ src/components/sign_up/hooks/useAuthAction.ts
