+++ src/views/sign_up/SignUp.entrypoint.tsx
@@ -1,5 +1,5 @@
-import AuthPageLayout from './AuthPageLayout';
-import SignUpCard from './SignUpCard';
+import AuthPageLayout from '../../components/sign_up/AuthPageLayout';
+import SignUpCard from '../../components/sign_up/SignUpCard';
 
 const SignUp = (): React.ReactElement => {
   return (
