--- src/views/sign_up/components/index.ts
@@ -1,9 +1,0 @@
-export { default as AuthCard } from './AuthCard';
-export { default as AuthCardHeader } from './AuthCardHeader';
-export {
-  ConfirmPasswordField,
-  EmailField,
-  PasswordField,
-} from './AuthFormFields';
-export { default as AuthSubmitButton } from './AuthSubmitButton';
-export { default as AuthSuccessMessage } from './AuthSuccessMessage';
